cpal/requirements.txt
# Core API framework
fastapi==0.111.0
uvicorn

# AWS Claude + Titan support
boto3
botocore
langchain-aws==0.2.27

# LangGraph + LangChain
langgraph==0.5.0
langgraph-prebuilt==0.5.1
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.27
langchain-community==0.3.26

# Vector store
qdrant-client
langchain-qdrant==0.2.0

# Evaluation
ragas==0.1.6

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.10.1
pydantic-settings    
